usuarios/adapters.py
from allauth.socialaccount.adapter import DefaultSocialAccountAdapter


class SocialAccountAdapter(DefaultSocialAccountAdapter):
    def populate_user(self, request, sociallogin, data):
        user = super().populate_user(request, sociallogin, data)
        # Usa a parte antes do @ do email como username
        if user.email:
            user.username = user.email.split('@')[0]
        return user