requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==6.0.3
django-allauth==65.15.0
django-environ==0.13.0
djangorestframework==3.17.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
qrcode==8.2
requests==2.33.0
sqlparse==0.5.5
urllib3==2.6.3
